requirements.txt
# YACLib Enhanced - Python Dependencies

# Core
requests>=2.32.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20
websockets>=13.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Image Processing
Pillow>=11.0.0

# Archive Support
rarfile>=4.2
py7zr>=0.22.0

# Performance (optional but recommended)
xxhash>=3.0.0  # 10x faster hashing for duplicate detection

# Utilities
python-dateutil>=2.9.0
humanize>=4.11.0

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.10.0
ruff>=0.8.0
mypy>=1.13.0

# Optional: Enhanced features
redis>=5.2.0  # For caching
aiofiles>=24.1.0  # Async file operations
watchdog>=6.0.0  # File system monitoring
